+++ dpc-go/dpc-attribution/src/repository/implementer_repository.go
@@ -5,13 +5,14 @@ import (
 	"database/sql"
 	"encoding/json"
 	"fmt"
-	"github.com/CMSgov/dpc/attribution/model"
+
+	"github.com/CMSgov/dpc/attribution/model/v2"
 	"github.com/huandu/go-sqlbuilder"
 )
 
 // ImplementerRepo is an interface for test mocking purposes
 type ImplementerRepo interface {
-	Insert(ctx context.Context, body []byte) (*model.Implementer, error)
+	Insert(ctx context.Context, body []byte) (*v2.Implementer, error)
 }
 
 // ImplementerRepository is a struct that defines what the repository has
@@ -27,15 +28,15 @@ func NewImplementerRepo(db *sql.DB) *ImplementerRepository {
 }
 
 // FindByID function that searches the database for the Implementer that matches the id
-func (or *ImplementerRepository) FindByID(ctx context.Context, id string) (*model.Implementer, error) {
+func (or *ImplementerRepository) FindByID(ctx context.Context, id string) (*v2.Implementer, error) {
 	sb := sqlFlavor.NewSelectBuilder()
 	sb.Select("id", "name", "created_at", "updated_at", "deleted_at")
 	sb.From("Implementer")
 	sb.Where(sb.Equal("id", id))
 	q, args := sb.Build()
 
-	Implementer := new(model.Implementer)
-	ImplementerStruct := sqlbuilder.NewStruct(new(model.Implementer)).For(sqlFlavor)
+	Implementer := new(v2.Implementer)
+	ImplementerStruct := sqlbuilder.NewStruct(new(v2.Implementer)).For(sqlFlavor)
 	if err := or.db.QueryRowContext(ctx, q, args...).Scan(ImplementerStruct.Addr(&Implementer)...); err != nil {
 		return nil, err
 	}
@@ -43,8 +44,8 @@ func (or *ImplementerRepository) FindByID(ctx context.Context, id string) (*mode
 }
 
 // Insert function that saves the Implementer model into the database and returns the model.Implementer
-func (or *ImplementerRepository) Insert(ctx context.Context, body []byte) (*model.Implementer, error) {
-	var ImplementerModel model.Implementer
+func (or *ImplementerRepository) Insert(ctx context.Context, body []byte) (*v2.Implementer, error) {
+	var ImplementerModel v2.Implementer
 	if err := json.Unmarshal(body, &ImplementerModel); err != nil {
 		return nil, err
 	}
@@ -61,8 +62,8 @@ func (or *ImplementerRepository) Insert(ctx context.Context, body []byte) (*mode
 
 	q, args := ib.Build()
 
-	Implementer := new(model.Implementer)
-	ImplementerStruct := sqlbuilder.NewStruct(new(model.Implementer)).For(sqlFlavor)
+	Implementer := new(v2.Implementer)
+	ImplementerStruct := sqlbuilder.NewStruct(new(v2.Implementer)).For(sqlFlavor)
 	if err := or.db.QueryRowContext(ctx, q, args...).Scan(ImplementerStruct.Addr(&Implementer)...); err != nil {
 		return nil, err
 	}
